+++ lib/portfolio/widgets/coin_list.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../../../../domain/entities/coin.dart';
+import '../model/coin.dart';
 import 'coin_details.dart';
 
 class CoinList extends StatelessWidget {
